requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.2.2
python-dotenv==1.0.0
openai==1.8.0